refactor/configd: move domain entities to own packages
