+++ configd/src/application/list_schemas.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 use std::sync::Arc;
 
-use crate::domain::{Error, SchemaRepository};
+use crate::domain::{errors::Error, schemas::SchemaRepository};
 
 #[derive(Deserialize)]
 pub struct ListSchemasCommand {
