+++ configd/src/application/update_config.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 use std::sync::Arc;
 
-use crate::domain::{Error, Id, Password, SchemaRepository};
+use crate::domain::{configs::Password, errors::Error, schemas::SchemaRepository, shared::Id};
 
 #[derive(Deserialize)]
 pub struct UpdateConfigCommand {
