+++ configd/src/domain/configs/config.rs
@@ -1,7 +1,12 @@
 use chrono::Duration;
 use core_lib::models::{Timestamps, Version};
 
-use crate::domain::{Access, Error, Id, Password, Value};
+use crate::domain::{
+    configs::{Access, Password},
+    errors::Error,
+    shared::Id,
+    values::Value,
+};
 
 #[derive(Debug, Clone)]
 pub struct Config {
