+++ configd/src/domain/configs/mod.rs
@@ -1,0 +1,7 @@
+mod access;
+mod config;
+mod password;
+
+pub use access::*;
+pub use config::*;
+pub use password::*;
