+++ configd/src/domain/mod.rs
@@ -1,27 +1,6 @@
-mod access;
-mod config;
-mod diff;
-mod errors;
-mod events;
-mod id;
-mod interval;
-mod json_prop;
-mod page;
-mod password;
-mod prop;
-mod schema;
-mod value;
+pub mod configs;
+pub mod schemas;
+pub mod shared;
+pub mod values;
 
-pub use access::*;
-pub use config::*;
-pub use diff::*;
-pub use errors::*;
-pub use events::*;
-pub use id::*;
-pub use interval::*;
-pub use json_prop::*;
-pub use page::*;
-pub use password::*;
-pub use prop::*;
-pub use schema::*;
-pub use value::*;
+pub mod errors;
