+++ configd/src/domain/schemas/events.rs
