+++ configd/src/domain/schemas/mod.rs
@@ -1,0 +1,5 @@
+mod events;
+mod schema;
+
+pub use events::*;
+pub use schema::*;
