+++ configd/src/domain/shared/mod.rs
@@ -1,0 +1,5 @@
+mod id;
+mod page;
+
+pub use id::*;
+pub use page::*;
