+++ configd/src/domain/values/mod.rs
@@ -1,0 +1,11 @@
+mod diff;
+mod interval;
+mod json_prop;
+mod prop;
+mod value;
+
+pub use diff::*;
+pub use interval::*;
+pub use json_prop::*;
+pub use prop::*;
+pub use value::*;
